# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /lists:
    get:
      operationId: listLists
      tags: [lists]
      summary: List the user's lists
      description: Returns the authenticated user's lists in manual order.
      parameters:
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: The user's lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/List"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createList
      tags: [lists]
      summary: Create a list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListRequest"
      responses:
        "201":
          description: The created list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /lists/{listId}:
    parameters:
      - $ref: "#/components/parameters/ListId"
    patch:
      operationId: updateList
      tags: [lists]
      summary: Update a list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListRequest"
      responses:
        "200":
          description: The updated list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      operationId: deleteList
      tags: [lists]
      summary: Delete a list
      description: >-
        Soft-deletes the list by archiving it and cascades every task it holds
        to deleted. The response reports how many tasks were affected.
      responses:
        "200":
          description: The archived list and its cascade count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteListResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /lists/{listId}/restore:
    parameters:
      - $ref: "#/components/parameters/ListId"
    post:
      operationId: restoreList
      tags: [lists]
      summary: Restore an archived list
      description: >-
        Un-archives the list and restores the tasks that were deleted by the
        same cascade, matched on deletedAt.
      responses:
        "200":
          description: The restored list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ListId:
      name: listId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: No such resource for this user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The request conflicts with the resource's current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: The request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: >-
        A machine-readable code plus copy that is safe to show the user.
        Per Principle III, no message assigns blame.
      properties:
        code:
          $ref: "#/components/schemas/ProblemCode"
        title:
          type: string
        detail:
          type: string
        fieldErrors:
          type: object
          description: Field name to message, for rendering inline beneath the field.
          additionalProperties:
            type: string
      required:
        - code
        - title
    ProblemCode:
      type: string
      enum:
        - Unauthorized
        - NotFound
        - DuplicateListName
        - InvalidTaskTransition
        - TaskAlreadyCompleted
        - UnknownTaskType
        - ValidationError
        - InternalError
...
    List:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        position:
          type: integer
        isArchived:
          type: boolean
        taskCount:
          type: integer
          description: Number of non-deleted tasks in the list.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, position, isArchived, taskCount, createdAt, updatedAt]
    DeleteListResult:
      type: object
      properties:
        list:
          $ref: "#/components/schemas/List"
        deletedTaskCount:
          type: integer
      required: [list, deletedTaskCount]
...
    CreateListRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        position:
          type: integer
      required: [name]
    UpdateListRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        position:
          type: integer
        isArchived:
          type: boolean
...