# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemCode } from "./problemCode";
import type { ProblemFieldErrors } from "./problemFieldErrors";

/**
 * A machine-readable code plus copy that is safe to show the user. Per Principle III, no message assigns blame.
 */
export interface Problem {
  code: ProblemCode;
  title: string;
  detail?: string;
  /** Field name to message, for rendering inline beneath the field. */
  fieldErrors?: ProblemFieldErrors;
}